+++ src/components/content/observations/helpers/base.tsx
@@ -1,3 +1,4 @@
+import cx from "classnames";
 import {
   ResourceTable,
   ResourceTableProps,
@@ -6,19 +7,19 @@ import {
   ResourceTableActions,
   ResourceTableActionsProps,
 } from "../../resource/resource-table-actions";
-import { patientObservationsColumns } from "./columns";
+import { diagnosticReportColumns } from "./columns";
 import { defaultObservationSort, observationSortOptions } from "./sorts";
 import { useObservationsDetailsDrawer } from "@/components/content/observations/helpers/drawer";
 import {
-  defaultObservationFilters,
-  diagnosticReportFilters,
+  defaultFilters,
+  filters,
 } from "@/components/content/observations/helpers/filters";
 import { DiagnosticReportModel } from "@/fhir/models";
 import { useFilteredSortedData } from "@/hooks/use-filtered-sorted-data";
 
 export type PatientObservationsTableProps = {
   action?: ResourceTableActionsProps<DiagnosticReportModel>["action"];
-  className?: string;
+  className?: cx.Argument;
   query: { data?: DiagnosticReportModel[]; isLoading: boolean };
   rowActions?: ResourceTableProps<DiagnosticReportModel>["rowActions"];
 };
@@ -33,16 +34,16 @@ export const PatientObservationsBase = ({
   const { data, setFilters, setSort } = useFilteredSortedData({
     defaultSort: defaultObservationSort,
     records: query.data,
-    defaultFilters: defaultObservationFilters,
+    defaultFilters,
   });
 
   return (
-    <div className={className}>
+    <div className={cx(className)}>
       <ResourceTableActions
         filterOptions={{
           onChange: setFilters,
-          defaultState: defaultObservationFilters,
-          filters: diagnosticReportFilters(query.data ?? []),
+          defaultState: defaultFilters,
+          filters: filters(query.data ?? []),
         }}
         sortOptions={{
           defaultSort: defaultObservationSort,
@@ -54,7 +55,7 @@ export const PatientObservationsBase = ({
 
       <ResourceTable
         className="ctw-patient-observations"
-        columns={patientObservationsColumns}
+        columns={diagnosticReportColumns}
         data={data}
         emptyMessage="There are no observation records available."
         isLoading={query.isLoading}
