+++ src/components/content/observations/helpers/columns.tsx
@@ -3,77 +3,43 @@ import { TableColumn } from "@/components/core/table/table-helpers";
 import { DiagnosticReportModel, ObservationModel } from "@/fhir/models";
 import { capitalize, startCase } from "@/utils/nodash";
 
-function statusToColor(
-  status: fhir4.DiagnosticReport["status"] | fhir4.Observation["status"]
-) {
-  switch (status) {
-    case "final":
-    case "amended":
-    case "appended":
-    case "corrected":
-      return "ctw-text-success-main";
-    case "cancelled":
-    case "entered-in-error":
-      return "ctw-text-caution-main";
-    case "registered":
-    case "partial":
-    case "preliminary":
-    case "unknown":
-    default:
-      return "ctw-text-content-lighter";
-  }
-}
-
-export const patientObservationsColumns: TableColumn<DiagnosticReportModel>[] =
-  [
-    {
-      title: "Report",
-      widthPercent: 30,
-      minWidth: 320,
-      render: (model) => (
-        <div>
-          <div className="ctw-medium-text">{model.displayName}</div>
+export const diagnosticReportColumns: TableColumn<DiagnosticReportModel>[] = [
+  {
+    title: "Report",
+    widthPercent: 30,
+    minWidth: 320,
+    dataIndex: "displayName",
+  },
+  {
+    title: "Category",
+    dataIndex: "category",
+  },
+  {
+    title: "Started On",
+    dataIndex: "effectiveStart",
+  },
+  {
+    title: "Status",
+    render: (model) => (
+      <div className="ctw-flex">
+        <div
+          className={cx("ctw-text-md", statusToColor(model.resource.status))}
+        >
+          &bull;
         </div>
-      ),
-    },
-    {
-      title: "Category",
-      render: (model) => <div>{model.category}</div>,
-    },
-    {
-      title: "Started On",
-      render: (model) => (
-        <div>
-          <div className="ctw-font-normal">
-            <div>{model.effectiveStart}</div>
-          </div>
+        <div className="ctw-ml-1">
+          <div>{capitalize(model.resource.status)}</div>
         </div>
-      ),
-    },
-    {
-      title: "Status",
-      render: (model) => (
-        <div className="ctw-flex">
-          <div
-            className={cx("ctw-text-md", statusToColor(model.resource.status))}
-          >
-            &bull;
-          </div>
-          <div className="ctw-ml-1">
-            <div>{capitalize(model.resource.status)}</div>
-          </div>
-        </div>
-      ),
-    },
-    {
-      title: "Observations",
-      render: (model) => (
-        <div className="ctw-flex ctw-justify-center">
-          <div>{model.results.length}</div>
-        </div>
-      ),
-    },
-  ];
+      </div>
+    ),
+  },
+  {
+    title: "Observations",
+    render: (model) => (
+      <div className="ctw-text-center">{model.results.length}</div>
+    ),
+  },
+];
 
 export const observationsColumns: TableColumn<ObservationModel>[] = [
   {
@@ -118,10 +84,31 @@ export const observationsColumns: TableColumn<ObservationModel>[] = [
   {
     title: "Notes",
     render: (model) =>
-      model.note && (
+      model.notes && (
         <blockquote className="ctw-mx-3 ctw-my-1 ctw-rounded ctw-bg-bg-lighter ctw-p-1.5 ctw-text-xs ctw-font-light ctw-italic">
-          {model.note}
+          {model.notes}
         </blockquote>
       ),
   },
 ];
+
+function statusToColor(
+  status: fhir4.DiagnosticReport["status"] | fhir4.Observation["status"]
+) {
+  switch (status) {
+    case "final":
+    case "amended":
+    case "appended":
+    case "corrected":
+      return "ctw-text-success-main";
+    case "cancelled":
+    case "entered-in-error":
+      return "ctw-text-caution-main";
+    case "registered":
+    case "partial":
+    case "preliminary":
+    case "unknown":
+    default:
+      return "ctw-text-content-lighter";
+  }
+}
