+++ src/components/content/observations/patient-observations-outside-badge.tsx
@@ -1,0 +1,14 @@
+import { Badge } from "@/components/core/badge";
+import { usePatientDiagnosticReportsOutside } from "@/fhir/diagnostic-report";
+
+export const PatientObservationsOutsideBadge = () => {
+  const { data = [] } = usePatientDiagnosticReportsOutside();
+
+  return data.length ? (
+    <Badge
+      color="notification"
+      text={data.length.toString()}
+      className="ctw-h-5"
+    />
+  ) : null;
+};
