+++ src/components/content/observations/patient-observations-outside.tsx
@@ -1,6 +1,5 @@
 import cx from "classnames";
 import { PatientObservationsBase } from "@/components/content/observations/helpers/base";
-import { Badge } from "@/components/core/badge";
 import { withErrorBoundary } from "@/components/core/error-boundary";
 import { usePatientDiagnosticReportsOutside } from "@/fhir/diagnostic-report";
 
@@ -22,15 +21,3 @@ export const PatientObservationsOutside = withErrorBoundary(
   Component,
   "PatientObservationsOutside"
 );
-
-export const PatientObservationsOutsideBadge = () => {
-  const { data = [] } = usePatientDiagnosticReportsOutside();
-
-  return data.length ? (
-    <Badge
-      color="notification"
-      text={data.length.toString()}
-      className="ctw-h-5"
-    />
-  ) : null;
-};
