+++ src/components/content/observations/patient-observations-profile.stories.tsx
@@ -1,15 +1,15 @@
 import { Meta, StoryObj } from "@storybook/react";
 import {
-  PatientObservations,
-  PatientObservationsProps as Props,
-} from "@/components/content/observations/patient-observations";
+  PatientObservationsProfile,
+  PatientObservationsProfileProps as Props,
+} from "@/components/content/observations/patient-observations-profile";
 import { setupObservationMocks } from "@/components/content/observations/story-helpers/mocks/requests";
 import { CTWProvider } from "@/components/core/providers/ctw-provider";
 import { PatientProvider } from "@/components/core/providers/patient-provider";
 import { SYSTEM_ZUS_UNIVERSAL_ID } from "@/fhir/system-urls";
 
 export default {
-  component: PatientObservations,
+  component: PatientObservationsProfile,
   tags: ["autodocs"],
   args: {},
   decorators: [
