+++ src/utils/query-keys.ts
@@ -4,13 +4,14 @@ export const QUERY_KEY_CARETEAM = "patientCareTeam";
 export const QUERY_KEY_CONDITION_HISTORY = "conditionHistory";
 export const QUERY_KEY_MEDICATION_HISTORY = "medication-history";
 export const QUERY_KEY_OTHER_PROVIDER_CONDITIONS = "otherProviderConditions";
-export const QUERY_KEY_OTHER_PROVIDER_DIAGNOSTIC_REPORTS = "otherDiagReports";
+export const QUERY_KEY_OTHER_PROVIDER_DIAGNOSTIC_REPORTS =
+  "otherProviderDiagnosticReports";
 export const QUERY_KEY_OTHER_PROVIDER_MEDICATIONS = "medications";
 export const QUERY_KEY_PATIENT = "patient";
 export const QUERY_KEY_PATIENT_ALLERGIES = "patientAllergies";
 export const QUERY_KEY_PATIENT_BUILDER_MEDICATIONS = "builderMedications";
 export const QUERY_KEY_PATIENT_CONDITIONS = "patientConditions";
-export const QUERY_KEY_PATIENT_DIAGNOSTIC_REPORTS = "diagReport";
+export const QUERY_KEY_PATIENT_DIAGNOSTIC_REPORTS = "diagnosticReports";
 export const QUERY_KEY_PATIENT_DOCUMENTS = "patientDocuments";
 export const QUERY_KEY_PATIENT_ENCOUNTERS = "patientEncounters";
 export const QUERY_KEY_PATIENT_HISTORY_DETAILS = "patientHistoryDetails";
